clean up TerraFabricPlugin
